refactor: move routing-related code to route.go
