+++ internal/delivery/http/route_boards.go
@@ -4,52 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
-	"strings"
 
 	"github.com/PichuChen/go-bbs"
 )
-
-// routeBoards is the handler for `/v1/boards`
-func (delivery *httpDelivery) routeBoards(w http.ResponseWriter, r *http.Request) {
-	delivery.logger.Debugf("routeBoards: %v", r)
-
-	// TODO: Check IP Flowspeed
-	if r.Method == "GET" {
-		delivery.getBoards(w, r)
-		return
-	}
-
-}
-
-// getBoards is the handler for `/v1/boards` with GET method
-func (delivery *httpDelivery) getBoards(w http.ResponseWriter, r *http.Request) {
-	delivery.logger.Debugf("getBoards: %v", r)
-	boardId, item, filename, err := delivery.parseBoardPath(r.URL.Path)
-	if boardId == "" {
-		delivery.getBoardList(w, r)
-		return
-	}
-	// get single board
-	if item == "information" {
-		delivery.getBoardInformation(w, r, boardId)
-		return
-	} else if item == "articles" {
-		if filename == "" {
-			delivery.getBoardArticles(w, r, boardId)
-		} else {
-			delivery.getBoardArticlesFile(w, r, boardId, filename)
-		}
-		return
-	} else if item == "treasures" {
-		delivery.getBoardTreasures(w, r, boardId)
-		return
-	}
-
-	// 404
-	w.WriteHeader(http.StatusNotFound)
-
-	delivery.logger.Noticef("board id: %v not exist but be queried, info: %v err: %v", boardId, item, err)
-}
 
 func (delivery *httpDelivery) getBoardList(w http.ResponseWriter, r *http.Request) {
 	delivery.logger.Debugf("getBoardList: %v", r)
@@ -138,39 +95,4 @@ func marshalBoardHeader(b bbs.BoardRecord) map[string]interface{} {
 		ret["type"] = "board"
 	}
 	return ret
-
-}
-
-func shouldShowOnUserLevel(b bbs.BoardRecord, u string) bool {
-	// TODO: Get user Level
-	return true
-}
-
-// parseBoardPath covert url path from /v1/boards/SYSOP/article to
-// {SYSOP, article) or /v1/boards to {,}
-func (delivery *httpDelivery) parseBoardPath(path string) (boardId string, item string, filename string, err error) {
-	pathSegment := strings.Split(path, "/")
-
-	if len(pathSegment) >= 6 {
-		// /{{version}}/boards/{{class_id}}/{{item}}/{{filename}}
-		boardId = pathSegment[3]
-		item = pathSegment[4]
-		filename = pathSegment[5]
-		return
-	} else if len(pathSegment) == 5 {
-		// /{{version}}/boards/{{class_id}}/{{item}}
-		boardId = pathSegment[3]
-		item = pathSegment[4]
-		return
-	} else if len(pathSegment) == 4 {
-		// /{{version}}/boards/{{class_id}}
-		boardId = pathSegment[3]
-		return
-	} else if len(pathSegment) == 3 {
-		// /{{version}}/boards
-		// Should not be reach...
-		return
-	}
-	delivery.logger.Warningf("parseBoardPath got malform path: %v", path)
-	return
 }
