+++ internal/delivery/http/route_boards_treasures.go
@@ -54,34 +54,3 @@ func (delivery *httpDelivery) getBoardTreasuresFile(w http.ResponseWriter, r *ht
 
 	w.WriteHeader(http.StatusNotImplemented)
 }
-
-// parseBoardTreasurePath parse covert url path from /v1/boards/SYSOP/article to
-// {SYSOP, article) or /v1/boards to {,}
-func (delivery *httpDelivery) parseBoardTreasurePath(path string) (boardId string, treasuresId []string, filename string, err error) {
-	pathSegment := strings.Split(path, "/")
-
-	if len(pathSegment) == 6 {
-		// /{{version}}/boards/{{board_id}}/treasures/articles
-		boardId = pathSegment[3]
-		treasuresId = []string{}
-		filename = ""
-		return
-	} else if len(pathSegment) >= 7 {
-		// /{{version}}/boards/{{board_id}}/treasures/{{treasures_id ... }}/articles
-		// or
-		// /{{version}}/boards/{{board_id}}/treasures/{{treasures_id ... }}/articles/{{filename}}
-		boardId = pathSegment[3]
-		if pathSegment[len(pathSegment)-1] == "articles" {
-			treasuresId = pathSegment[5 : len(pathSegment)-1]
-			filename = ""
-		} else {
-			treasuresId = pathSegment[5 : len(pathSegment)-2]
-			filename = pathSegment[len(pathSegment)-1]
-		}
-		return
-	}
-	// should not be reached
-	delivery.logger.Warningf("parseBoardTreasurePath got malform path: %v", path)
-	return
-
-}
